requirements.txt
streamlit>=1.28.0
gspread>=5.12.0
oauth2client>=4.1.3
PyMuPDF>=1.23.0
google-generativeai>=0.3.0
python-dateutil>=2.8.2